Fetch content for the footer from the locale files
The content in the footer is now fetched from the locale files so we
can alternate the footer’s content per language. We allow setting plain
html via React’s `dangerouslySetInnerHTML` so we can use links as
footer content.

It also introduces a new convention: suffix a locale key with `_html` to
indicate that the specified key contains html content.