Add basic idp switcher to the header
This allows the user to select a different idp from the provided
options. It makes sure that the api is set to the different idp before
refreshing the page and reloading all data with the new idp.

It includes some basic styling but it’s also marked with an `ugly`
class, so the designer can take a look at it.